refactor: manage patients for doctor
